requirements.txt
Django==4.0
djangorestframework==3.13.1
mysqlclient==2.1.0
django-mysql==4.4.0
django-cors-headers==3.11.0
pika==1.2.0